Leave usergroup later and use a more performant way to query privilege
Since implementing usergroup requires lots of extra work to do and the
use case is not very obvious now, I'll leave it later. Meanwhile,
isroot and contest_owner column are retained to avoid adding to much
entries in privilege table. I also move the methods from concerns to a
helper.