Switch user auth to session_based
A session model is added, with ip and expire date. The expire date is updated every time a user hits the website if remerber me is checked when logged in.

Rails session mechanism can only has fixed expire time, so I use database to store session. Of course cookies is always an option, but database gives us more flexibility